Beginning of javaapi
