+++ javaapi/.classpath
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<classpath>
+	<classpathentry kind="src" path=""/>
+	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER">
+		<attributes>
+			<attribute name="module" value="true"/>
+		</attributes>
+	</classpathentry>
+	<classpathentry kind="lib" path="/run/media/guilherme/2D4B3415324E94B7/guilherme/Google Drive/Projetos/eyesofthundera/javaapi/newssites/json-simple-1.1.1.jar"/>
+	<classpathentry kind="output" path=""/>
+</classpath>
