--- javaapi/Requests.java
@@ -1,108 +1,0 @@
-import java.io.BufferedReader;
-import java.io.DataOutputStream;
-import java.io.InputStreamReader;
-import java.net.HttpURLConnection;
-import java.net.URL;
-import org.json.simple.JSONObject;
-import javax.net.ssl.HttpsURLConnection;
-import javax.net.ssl.HttpPost;
-
-public class Requests {
-
-	private final String USER_AGENT = "Mozilla/5.0";
-
-	public static void main(String[] args) throws Exception {
-
-		Requests http = new Requests();
-
-        JSONObject obj = new JSONObject();
-
-        obj.put("keyword", "json");
-        obj.put("blacklist", new Boolean(true));
-
-        http.sendGet("http://localhost:3000/keywords/add",obj);
-
-	}
-
-	// HTTP GET request
-	public static void sendGet(String url, JSONObject jsondata ) throws Exception {
-		
-		URL obj = new URL(url);
-        HttpClient httpClient = HttpClientBuilder.create().build(); //Use this instead 
-		
-        //try {
-            HttpPost request = new HttpPost(url);
-            StringEntity params = new StringEntity(jsondata);
-            request.addHeader("content-type", "application/x-www-form-urlencoded");
-            request.setEntity(params);
-            HttpResponse response = httpClient.execute(request);
-            System.out.println(reponse.body());
-        //}
-        
-        /*// optional default is GET
-		con.setRequestMethod("GET");
-
-		//add request header
-		con.setRequestProperty("User-Agent", USER_AGENT);
-
-		int responseCode = con.getResponseCode();
-		System.out.println("\nSending 'GET' request to URL : " + url);
-		System.out.println("Response Code : " + responseCode);
-
-		BufferedReader in = new BufferedReader(
-		        new InputStreamReader(con.getInputStream()));
-		String inputLine;
-		StringBuffer response = new StringBuffer();
-
-		while ((inputLine = in.readLine()) != null) {
-			response.append(inputLine);
-		}
-		in.close();
-
-		//print result
-		System.out.println(response.toString());*/
-
-	}
-	
-	// HTTP POST request
-	private void sendPost() throws Exception {
-
-		String url = "https://selfsolve.apple.com/wcResults.do";
-		URL obj = new URL(url);
-		HttpsURLConnection con = (HttpsURLConnection) obj.openConnection();
-
-		//add reuqest header
-		con.setRequestMethod("POST");
-		con.setRequestProperty("User-Agent", USER_AGENT);
-		con.setRequestProperty("Accept-Language", "en-US,en;q=0.5");
-
-		String urlParameters = "sn=C02G8416DRJM&cn=&locale=&caller=&num=12345";
-		
-		// Send post request
-		con.setDoOutput(true);
-		DataOutputStream wr = new DataOutputStream(con.getOutputStream());
-		wr.writeBytes(urlParameters);
-		wr.flush();
-		wr.close();
-
-		int responseCode = con.getResponseCode();
-		System.out.println("\nSending 'POST' request to URL : " + url);
-		System.out.println("Post parameters : " + urlParameters);
-		System.out.println("Response Code : " + responseCode);
-
-		BufferedReader in = new BufferedReader(
-		        new InputStreamReader(con.getInputStream()));
-		String inputLine;
-		StringBuffer response = new StringBuffer();
-
-		while ((inputLine = in.readLine()) != null) {
-			response.append(inputLine);
-		}
-		in.close();
-		
-		//print result
-		System.out.println(response.toString());
-
-	}
-
-}
