+++ javaapi/newssites/.gitignore
@@ -1,0 +1,2 @@
+/Requests.class
+/HttpURLConnectionExample.class
